Long Distance Switching/Receiver Script/requirements.txt
pyFirmata
requests
urllib3
pyserial
